added role based authentication as well as ownership check to critical actions such as update and viewig, and deleting resourse

+++ src/main/java/sample/project/Controller/UserController.java
@@ -5,6 +5,10 @@ import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,15 +18,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import jakarta.validation.Valid;
+import sample.project.Auth.JwtService;
 import sample.project.DTO.request.LoginRequest;
 import sample.project.DTO.request.RegisterRequest;
 import sample.project.DTO.response.LoginResponse;
 import sample.project.DTO.response.RegisterResponse;
 import sample.project.DTO.response.UserResponse;
 import sample.project.DTO.response.UserResponseList;
+import sample.project.Model.User;
 import sample.project.Service.UserService;
 
 @RestController
@@ -36,7 +43,6 @@ public class UserController {
     }
 
     @PostMapping("/public")
-    // @Authorization("admin")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<RegisterResponse> createUser(@Valid @RequestBody RegisterRequest req) {
         RegisterResponse response = userService.createUser(req);
@@ -55,13 +61,18 @@ public class UserController {
 
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<UserResponse> getUser(@PathVariable Long id) {
-        UserResponse responseList = userService.getUser(id);
-        return ResponseEntity.ok().body(responseList);
+    public ResponseEntity<UserResponse> getUser(@PathVariable Long id, @AuthenticationPrincipal User currentUser) {
+
+        if (!currentUser.getId().equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
+        UserResponse response = userService.getUser(id);
+        return ResponseEntity.ok().body(response);
     }
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<UserResponseList> getUsers() {
 
         UserResponseList response = userService.getAllUser();
@@ -71,7 +82,14 @@ public class UserController {
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<UserResponse> updateUser(@PathVariable Long id, @RequestBody RegisterRequest req) {
+    @PreAuthorize("hasRole('USER')")
+    public ResponseEntity<UserResponse> updateUser(@PathVariable Long id, @RequestBody RegisterRequest req,
+            @AuthenticationPrincipal User currentUser) {
+        if (!currentUser.getId().equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+
+        }
+
         UserResponse response = userService.updateUser(req, id);
         return ResponseEntity.ok().body(response);
 
@@ -79,7 +97,13 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Long id) {
+    @PreAuthorize("hasRole('USER')")
+    public void deleteUser(@PathVariable Long id, @AuthenticationPrincipal User currentUser) {
+
+        if (!currentUser.getId().equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can only delete your own account");
+        }
+
         userService.deleteUser(id);
     }
